linter corrections

+++ jobs/rocoto/earc.sh
@@ -56,11 +56,11 @@ if [ "${status}" -ne 0 ]; then
    exit "${status}"
 fi
 
-if [ $DO_EFSOI = "YES" ]; then
-   $HOMEgfs/ush/hpssarch_gen.sh efsoigdas 
+if [ "${DO_EFSOI}" = "YES" ]; then
+   "${HOMEgfs}/ush/hpssarch_gen.sh efsoigdas" 
    status=$?
    if [ "${status}" -ne 0 ]; then
-      echo "$HOMEgfs/ush/hpssarch_gen_EFSOI.sh enkf${CDUMP} failed, ABORT!"
+      echo "${HOMEgfs}/ush/hpssarch_gen_EFSOI.sh enkf${CDUMP} failed, ABORT!"
       exit "${status}"
    fi
 fi
@@ -252,14 +252,14 @@ if [ $ENSGRP -eq 0 ]; then
 	done
 
        if [ "${DO_EFSOI}" = "YES" ]; then
-            COMIN_ENS="$i{ROTDIR}/efsoigdas.${gPDY}/${gcyc}/${COMPONENT}"
+            COMIN_ENS="${ROTDIR}/efsoigdas.${gPDY}/${gcyc}/${COMPONENT}"
             if [ -d "${COMIN_ENS}" ] ; then
                 rm -rf "${COMIN_ENS}/*f012*nc"
                 rm -rf "${COMIN_ENS}/*f018*nc"
                 for imem in $(seq 1 "${NMEM_ENKF}"); do
                     memchar="mem"$(printf %03i "${imem}")
                     for file in $(ls "${COMIN_ENS}/${memchar}" |grep -v atmf024); do
-                       rm -rf "${COMIN_ENS}/${memchar}/${file}"
+                       rm -rf "${COMIN_ENS:?}/${memchar}/${file}"
                     done
                 done
    	    fi
@@ -273,24 +273,24 @@ if [ $ENSGRP -eq 0 ]; then
     if [ "${DO_EFSOI}" = "YES" ]; then
         # Now do EFSOI - needs to be kept around longer
         # Start start and end dates to remove
-        GDATEEND=$($NDATE -${RMOLDEND_EFSOI:-36}  $CDATE)
-        GDATE=$($NDATE -${RMOLDSTD_ENKF:-120} $CDATE)
-        while [ $GDATE -le $GDATEEND ]; do
+        GDATEEND=$("${NDATE}" -${RMOLDEND_EFSOI:-36}  "${CDATE}")
+        GDATE=$("${NDATE}" -${RMOLDSTD_ENKF:-120} "${CDATE}")
+        while [ "${GDATE}" -le "${GDATEEND}" ]; do
     
-           gPDY=$(echo $GDATE | cut -c1-8)
-           gcyc=$(echo $GDATE | cut -c9-10)
+           gPDY=$(echo "${GDATE}" | cut -c1-8)
+           gcyc=$(echo "${GDATE}" | cut -c9-10)
     
-           COMIN_ENS="$ROTDIR/efsoigdas.$gPDY/$gcyc/$COMPONENT"
-             [[ -d $COMIN_ENS ]] && rm -rf $COMIN_ENS
+           COMIN_ENS="${ROTDIR}/efsoigdas.${gPDY}/${gcyc}/${COMPONENT}"
+             [[ -d "${COMIN_ENS}" ]] && rm -rf "${COMIN_ENS}"
     
            # Remove any empty directories
-           COMIN_ENS="$ROTDIR/efsoigdas.$gPDY/$COMPONENT"
-           if [ -d $COMIN_ENS ] ; then
-               [[ ! "$(ls -A $COMIN_ENS)" ]] && rm -rf $COMIN_ENS
+           COMIN_ENS="${ROTDIR}/efsoigdas.${gPDY}/${COMPONENT}"
+           if [ -d "${COMIN_ENS} ] ; then
+               [[ ! "$(ls -A ${COMIN_ENS})" ]] && rm -rf "${COMIN_ENS}"
            fi
            
            # Advance to next cycle
-           GDATE=$($NDATE +$assim_freq $GDATE)
+           GDATE=$("${NDATE}" +$assim_freq "${GDATE}")
            
        done
     fi # $DO_EFSOI = "YES"
